requirements.txt
alabaster==0.7.12
alembic==1.4.3
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.0
bidict==0.21.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
decouple==0.0.7
Discord-Webhooks==1.0.4
distlib==0.3.1
dj-database-url==0.5.0
# Editable install with no version control (Django==3.1.2)
-e /Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages
django-crispy-forms==1.10.0
django-environ==0.4.5
django-extensions==3.0.9
django-fontawesome-5==1.0.18
django-img==0.1.5
djangorestframework==3.12.2
docutils==0.16
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MailAlchemy==0.4
Flask-Migrate==2.5.3
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
idna==2.8
imagesize==1.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-client==0.0.1
mysqlclient==2.0.1
packaging==20.8
parso==0.7.1
pbr==5.5.1
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==4.0.0
python-http-client==3.3.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.1
python-socketio==5.0.4
python-utils==2.4.0
pytz==2020.1
recommonmark==0.7.1
requests==2.21.0
rope==0.18.0
schedule==0.6.0
selenium==3.141.0
sendgrid==6.4.7
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.22
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stevedore==3.3.0
toml==0.10.2
ujson==4.0.1
Unidecode==1.1.1
urllib3==1.24.3
virtualenv==20.0.35
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
yapf==0.30.0
